refactored domain entity events types and classes
