--- src/events/classes/baseDomainEntityOrAggregateEvent/index.ts
@@ -1,1 +1,0 @@
-export {BaseDomainEntityOrAggregateEvent} from './baseDomainEntityOrAggregateEvent';
