+++ src/events/classes/index.ts
@@ -2,8 +2,6 @@ export * from './baseDomainEntityConstructEvent';
 
 export * from './baseDomainEntityDeleteEvent';
 
-export * from './baseDomainEntityOrAggregateEvent';
-
 export * from './baseDomainEntityEvent';
 
 export * from './baseDomainEvent';
