+++ src/events/constants/eventType.ts
@@ -1,4 +1,3 @@
 export enum EDomainEntityEventType {
   ENTITY_EVENT = 'ENTITY_EVENT',
-  AGGREGATE_EVENT = 'AGGREGATE_EVENT',
 }
